Move DoFollows into the pkg module.
